backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML & Data Processing
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# Google Cloud
google-cloud-storage==2.10.0
google-cloud-secret-manager==2.16.4

# External APIs
requests==2.31.0
python-dotenv==1.0.0

# Date/Time & Holidays
python-dateutil==2.8.2
holidays==0.36

# Logging & Monitoring
python-json-logger==2.0.7

# CORS
fastapi-cors==0.0.6

# Testing (optional)
pytest==7.4.3
httpx==0.25.2
